+++ examples/read-indoor-navigation.js
@@ -1,9 +1,8 @@
 'use strict';
 
 /*
- * Here the BME280 is configured to run using the default options. The
- * defaults for oversampling and filtering are those recommended by the BME280
- * datasheet for indoor navigation.
+ * Here the BME280 is configured to run using oversampling and filtering
+ * options recommended for indoor navigation by the BME280 datasheet.
  */
 
 const bme280 = require('../');
@@ -12,7 +11,14 @@ const format = number => (Math.round(number * 100) / 100).toFixed(2);
 const delay = millis => new Promise(resolve => setTimeout(resolve, millis));
 
 const reportContinuous = async _ => {
-  const sensor = await bme280.open();
+  const sensor = await bme280.open({
+    i2cBusNumber: 1,
+    i2cBusAddress: 0x77,
+    humidityOversampling: bme280.OVERSAMPLE.X1,
+    pressureOversampling: bme280.OVERSAMPLE.X16,
+    temperatureOversampling: bme280.OVERSAMPLE.X2,
+    filterCoefficient: bme280.FILTER.F16
+  });
 
   for (let i = 1; i <= 250; ++i) {
     const reading = await sensor.read();