+++ History.md
@@ -2,6 +2,7 @@ Unpublished
 ===========
 
   * drop support for node.js 6 and start using await and async
+  * add support for oversampling, filtering and forced mode
 
 1.0.0 - Oct 03 2019
 ===================